Change index types to `u16` (#79)
* oracle: change the index type for `PriceFeed` to `u16`

* order: change the index type for `TradeData` to `u16`

* GLV: change the index type for `Glv` to `u16`

* treasury: change the index type for `TreasuryVaultConfig` to `u16`

+++ crates/gmsol/src/cli/glv.rs
@@ -179,7 +179,7 @@ struct GlvToken {
     glv_token: Option<Pubkey>,
     /// Index.
     #[arg(long)]
-    index: Option<u8>,
+    index: Option<u16>,
 }
 
 #[derive(clap::Args)]
